// web/src/api/generated/models/authStaffDetailResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * POS Cafe API
 * Backend API for Cafe Point of Sale System built with Vertical Slice Architecture and Idiomatic Go.
 * OpenAPI spec version: 1.0
 */

export interface AuthStaffDetailResponse {
  created_at?: string;
  display_name?: string;
  enabled?: boolean;
  id?: string;
  login_code?: string;
  roles?: string[];
}